+spr #518 enable autoconfiguration of akka streams in spring web
