+++ spring-web/src/test/java/akka/stream/alpakka/spring/web/DemoApplication.java
@@ -3,18 +3,9 @@
  */
 package akka.stream.alpakka.spring.web;
 
-import akka.actor.ActorSystem;
-import akka.event.Logging;
-import akka.stream.ActorMaterializer;
-import akka.stream.Materializer;
-import akka.stream.javadsl.AsPublisher;
-import org.springframework.beans.factory.annotation.Autowired;
+//#use
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.core.ReactiveAdapterRegistry;
-import org.springframework.util.Assert;
-
-import static org.springframework.core.ReactiveTypeDescriptor.multiValue;
 
 @SpringBootApplication
 public class DemoApplication {
@@ -24,3 +15,4 @@ public class DemoApplication {
   }
 
 }
+//#use
